refactor: new store.ts based on old vanilla.ts
