+++ src/core/Provider.ts
@@ -1,14 +1,16 @@
-import { createElement, useDebugValue, useRef } from 'react'
+import { createElement, useCallback, useDebugValue, useRef } from 'react'
 import type { PropsWithChildren } from 'react'
 import type { Atom, Scope } from './atom'
 import {
+  ScopeContainer,
   createScopeContainer,
   getScopeContext,
   isDevScopeContainer,
 } from './contexts'
 import type { ScopeContainerForDevelopment } from './contexts'
+import { DEV_GET_ATOM_STATE, DEV_GET_MOUNTED } from './store'
+import type { AtomState, Store } from './store'
 import { useMutableSource } from './useMutableSource'
-import type { AtomState, State } from './vanilla'
 
 export const Provider = ({
   initialValues,
@@ -18,10 +20,8 @@ export const Provider = ({
   initialValues?: Iterable<readonly [Atom<unknown>, unknown]>
   scope?: Scope
 }>) => {
-  const scopeContainerRef = useRef<ReturnType<
-    typeof createScopeContainer
-  > | null>(null)
-  if (scopeContainerRef.current === null) {
+  const scopeContainerRef = useRef<ScopeContainer>()
+  if (!scopeContainerRef.current) {
     // lazy initialization
     scopeContainerRef.current = createScopeContainer(initialValues)
   }
@@ -51,15 +51,15 @@ export const Provider = ({
 const atomToPrintable = (atom: Atom<unknown>) =>
   atom.debugLabel || atom.toString()
 
-const stateToPrintable = ([state, atoms]: [State, Atom<unknown>[]]) =>
+const stateToPrintable = ([store, atoms]: [Store, Atom<unknown>[]]) =>
   Object.fromEntries(
     atoms.flatMap((atom) => {
-      const mounted = state.m.get(atom)
+      const mounted = store[DEV_GET_MOUNTED]?.(atom)
       if (!mounted) {
         return []
       }
       const dependents = mounted.d
-      const atomState = state.a.get(atom) || ({} as AtomState)
+      const atomState = store[DEV_GET_ATOM_STATE]?.(atom) || ({} as AtomState)
       return [
         [
           atomToPrintable(atom),
@@ -72,30 +72,14 @@ const stateToPrintable = ([state, atoms]: [State, Atom<unknown>[]]) =>
     })
   )
 
-export const getDebugStateAndAtoms = ({
-  atoms,
-  state,
-}: {
-  atoms: Atom<unknown>[]
-  state: State
-}) => [state, atoms] as const
-
-export const subscribeDebugScopeContainer = (
-  { listeners }: { listeners: Set<() => void> },
-  callback: () => void
-) => {
-  listeners.add(callback)
-  return () => listeners.delete(callback)
-}
-
 // We keep a reference to the atoms in Provider's registeredAtoms in dev mode,
 // so atoms aren't garbage collected by the WeakMap of mounted atoms
 const useDebugState = (scopeContainer: ScopeContainerForDevelopment) => {
-  const debugMutableSource = scopeContainer[4]
-  const [state, atoms] = useMutableSource(
-    debugMutableSource,
-    getDebugStateAndAtoms,
-    subscribeDebugScopeContainer
+  const [store, , devMutableSource, devSubscribe] = scopeContainer
+  const atoms = useMutableSource(
+    devMutableSource,
+    useCallback((devContainer) => devContainer.atoms, []),
+    devSubscribe
   )
-  useDebugValue([state, atoms], stateToPrintable)
+  useDebugValue([store, atoms], stateToPrintable)
 }
