+++ src/core/contexts.ts
@@ -1,64 +1,50 @@
 import { createContext } from 'react'
 import type { Context } from 'react'
-import type { Atom, Scope, WritableAtom } from './atom'
+import type { Atom, Scope } from './atom'
+import { GET_VERSION, createStore } from './store'
 import { createMutableSource } from './useMutableSource'
-import { createState, flushPending, restoreAtoms, writeAtom } from './vanilla'
 
 const createScopeContainerForProduction = (
   initialValues?: Iterable<readonly [Atom<unknown>, unknown]>
 ) => {
-  const state = createState(initialValues)
-  const stateMutableSource = createMutableSource(state, () => state.v)
-  const commitCallback = () => flushPending(state)
-  const updateAtom = <Value, Update>(
-    atom: WritableAtom<Value, Update>,
-    update: Update
-  ) => writeAtom(state, atom, update)
-  const restore = (values: Iterable<readonly [Atom<unknown>, unknown]>) =>
-    restoreAtoms(state, values)
-  return [stateMutableSource, updateAtom, commitCallback, restore] as const
+  const store = createStore(initialValues)
+  const mutableSource = createMutableSource(store, store[GET_VERSION])
+  return [store, mutableSource] as const
 }
 
 const createScopeContainerForDevelopment = (
   initialValues?: Iterable<readonly [Atom<unknown>, unknown]>
 ) => {
+  let devVersion = 0
+  const devListeners = new Set<() => void>()
+  const devContainer = {
+    atoms: Array.from(initialValues ?? []).map(([a]) => a),
+  }
   const stateListener = (updatedAtom: Atom<unknown>, isNewAtom: boolean) => {
-    ++debugContainer.version
+    ++devVersion
     if (isNewAtom) {
       // FIXME memory leak
       // we should probably remove unmounted atoms eventually
-      debugContainer.atoms = [...debugContainer.atoms, updatedAtom]
+      devContainer.atoms = [...devContainer.atoms, updatedAtom]
     }
     Promise.resolve().then(() => {
-      debugContainer.listeners.forEach((listener) => listener())
+      devListeners.forEach((listener) => listener())
     })
   }
-  const state = createState(initialValues, stateListener)
-  const stateMutableSource = createMutableSource(state, () => state.v)
-  const commitCallback = () => flushPending(state)
-  const updateAtom = <Value, Update>(
-    atom: WritableAtom<Value, Update>,
-    update: Update
-  ) => writeAtom(state, atom, update)
-  const debugContainer = {
-    version: 0,
-    atoms: Array.from(initialValues ?? []).map(([a]) => a),
-    state,
-    listeners: new Set<() => void>(),
+  const store = createStore(initialValues, stateListener)
+  const mutableSource = createMutableSource(store, store[GET_VERSION])
+  const devMutableSource = createMutableSource(devContainer, () => devVersion)
+  const devSubscribe = (_: unknown, callback: () => void) => {
+    devListeners.add(callback)
+    return () => devListeners.delete(callback)
   }
-  const debugMutableSource = createMutableSource(
-    debugContainer,
-    () => debugContainer.version
-  )
-  const restore = (values: Iterable<readonly [Atom<unknown>, unknown]>) =>
-    restoreAtoms(state, values)
-  return [
-    stateMutableSource,
-    updateAtom,
-    commitCallback,
-    restore,
-    debugMutableSource,
-  ] as const
+  return [store, mutableSource, devMutableSource, devSubscribe] as const
+}
+
+export const isDevScopeContainer = (
+  scopeContainer: ScopeContainer
+): scopeContainer is ScopeContainerForDevelopment => {
+  return scopeContainer.length > 2
 }
 
 type ScopeContainerForProduction = ReturnType<
@@ -89,10 +75,4 @@ export const getScopeContext = (scope?: Scope) => {
     ScopeContextMap.set(scope, createContext(createScopeContainer()))
   }
   return ScopeContextMap.get(scope) as ScopeContext
-}
-
-export const isDevScopeContainer = (
-  store: ScopeContainer
-): store is ScopeContainerForDevelopment => {
-  return store.length > 4
 }
